+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Show error message when changelog file cannot be found, instead of crashing
 - Support Jira ticket keys with digits before the dash: K23-42
+- Support TYPO3 v11 and v13
 
 ### Changed